+++ com.chartboost.mediation/Runtime/Mediation/Utilities/Events/BannerAdEvents.cs
@@ -10,6 +10,8 @@ namespace Chartboost.Mediation.Utilities.Events
         Load = 0,
         Click = 1,
         RecordImpression = 2,
-        Drag = 3
+        BeginDrag = 3,
+        Drag = 4,
+        EndDrag = 5
     }
 }