+++ com.chartboost.mediation/Runtime/iOS/Plugins/iOS/Utilities/CBMAdStore.m
@@ -1,5 +1,7 @@
 #import "CBMAdStore.h"
 
+NSString* const CBMAdStoreTAG = @"CBMAdStore";
+
 @implementation CBMAdStore
 
 + (instancetype)sharedStore {
@@ -16,15 +18,18 @@
 - (void)storeAd:(id)ad {
     NSNumber *key = [NSNumber numberWithLong:(long)ad];
     [_adStore setObject:ad forKey:key];
+    [[CBLUnityLoggingBridge sharedLogger] logWithTag:CBMAdStoreTAG log:[NSString stringWithFormat:@"Stored ad with key: %@", key] logLevel:CBLLogLevelDebug];
 }
 
 - (void)storeAd:(id)ad withKey:(NSNumber *)key
 {
     [_adStore setObject:ad forKey:key];
+    [[CBLUnityLoggingBridge sharedLogger] logWithTag:CBMAdStoreTAG log:[NSString stringWithFormat:@"Stored ad with key: %@", key] logLevel:CBLLogLevelDebug];
 }
 
 - (void)releaseAd:(const void *)uniqueId {
     NSNumber *key = [NSNumber numberWithLong:(long)uniqueId];
     [_adStore removeObjectForKey:key];
+    [[CBLUnityLoggingBridge sharedLogger] logWithTag:CBMAdStoreTAG log:[NSString stringWithFormat:@"Removed ad from store with key: %@", key] logLevel:CBLLogLevelDebug];
 }
 @end